Make t_item_range_value a signed 64-bit value. Fixes #1153

+++ src/cov/cov-api.h
@@ -89,8 +89,8 @@ typedef enum {
 } cover_src_t;
 
 typedef struct {
-   uint32_t  min;
-   uint32_t  max;
+   int64_t min;
+   int64_t max;
 } cover_range_t;
 
 typedef struct _cover_item {
