+++ src/rt/cover.c
@@ -470,7 +470,7 @@ void _nvc_add_cover_item(jit_scalar_t *args)
    item->n_ranges = ffi_array_length(args[10].integer);
    item->ranges = xcalloc_array(item->n_ranges, sizeof(cover_range_t));
 
-   int32_t *ptr = (int32_t *)args[8].pointer;
+   int64_t *ptr = (int64_t *)args[8].pointer;
 
    for (int i = 0; i < item->n_ranges; i++) {
       item->ranges[i].min = *ptr++;
